requirements.txt
pandas>=1.3.0
requests>=2.25.0
scikit-learn>=1.0.0
imbalanced-learn>=0.9.0
tranco
python-whois
cryptography
requests
python-dotenv